Reset the title of tmux panes on creation
According to
https://manpages.debian.org/bookworm/tmux/tmux.1.en.html#NAMES_AND_TITLES:

> When a pane is first created, its title is the hostname.

Since the title of the pane is shown in the status bar, having it in
every pane is redundant. However, there is no way to customize this
behavior, so we reset the pane title instead. The variable `$TMUX` is
documented in
https://manpages.debian.org/bookworm/tmux/tmux.1.en.html#GLOBAL_AND_SESSION_ENVIRONMENT:

> tmux also initialises the TMUX variable with some internal information…

Not using the "-v" switch because we don’t differentiate between blank
and unset values.